Adopt a new design and move existing conversion code
